+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -20,6 +20,7 @@ CHANGELOG
 
  * added `AddValidatorInitializersPass`
  * added `AddConstraintValidatorsPass`
+ * added `ContainerConstraintValidatorFactory`
 
 3.2.0
 -----